Refactor Dp implementation and move to Density
Modified the Dp implementation from using a Float representation to using Double. This provides more
precision which is useful for handling unit conversions.

The Dp implementation was also moved to a new file, Density.kt, for better organization and readability.
The file Dp.kt was deleted and a new one was created in its place.

This change supports better handling of device-independent pixels in various sections of the code by
making the value more precise and appropriately grouped with related conversions in the Density.kt file.