Refactor GraphicsContext into a property of RenderThread
Removed GraphicsContext parameter from methods in Driver.kt to simplify the API for the client side. Instead, the GraphicsContext is accessed via a property in the RenderThread interface. This change reduces the complexity and redundancy of passing GraphicsContext as a method parameter across different areas of the application.


Merge-request: TINN-MR-31
Merged-by: Fabian Fahrenholz <fabianfahrenholz718@gmail.com>