Add requireAlive function to Destroyable interface
This commit introduces a new helper function, requireAlive, to the Destroyable interface. Its purpose is to provide a straightforward mechanism to verify that a Destroyable object has not been destroyed yet, hence is alive. This can be used to prevent operations on destroyed objects, improving safety across the system.


Merge-request: TINN-MR-33
Merged-by: Fabian Fahrenholz <fabianfahrenholz718@gmail.com>