Introduced Tracker class and simplified imports
In this commit, a new Tracker class was introduced in the `com.virusbear.tinn` package to handle tracking of Destroyable objects. This new class was then used to replace the previous approach of directly managing a mutable set of tracked objects within the Driver and GraphicsContext classes, leading to cleaner and more maintainable code. Additionally, for better readability, all related imports in the GraphicsContextGL class are consolidated into a single wildcard import.


Merge-request: TINN-MR-37
Merged-by: Fabian Fahrenholz <fabianfahrenholz718@gmail.com>